+++ src/renderer/canvas/index.js
@@ -12,7 +12,6 @@ module.exports = {
 
     BlitImage: require('./utils/BlitImage'),
     CanvasRenderer: require('./CanvasRenderer'),
-    DrawImage: require('./utils/DrawImage'),
     GetBlendModes: require('./utils/GetBlendModes'),
     SetTransform: require('./utils/SetTransform')